Fix TypeScript errors

+++ pages/dashboard.tsx
@@ -1,34 +1,54 @@
-import { auth, db } from '@/utils/firebase';
-import { useAuthState } from 'react-firebase-hooks/auth';
-import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
-import { collection, deleteDoc, doc, onSnapshot, query, where } from 'firebase/firestore';
-import Message from '@/components/message';
-import {BsTrash2Fill} from 'react-icons/bs';
-import {AiFillEdit} from 'react-icons/ai';
-import Link from 'next/link';
+import { auth, db } from "@/utils/firebase";
+import { useAuthState } from "react-firebase-hooks/auth";
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
+import {
+  collection,
+  deleteDoc,
+  doc,
+  onSnapshot,
+  query,
+  where,
+} from "firebase/firestore";
+import Message from "@/components/message";
+import { BsTrash2Fill } from "react-icons/bs";
+import { AiFillEdit } from "react-icons/ai";
+import Link from "next/link";
 
-export default function Dashboard(){
+interface Post {
+  comments?: string[];
+  description?: string;
+  id: string;
+  timestamp?: string;
+}
+
+export default function Dashboard() {
   const route = useRouter();
-  const [posts, setPosts] = useState([]);
+  const [posts, setPosts] = useState<Post[]>([]);
   const [user, loading] = useAuthState(auth);
   //See if user is Logged
   const getData = async () => {
-    if(loading) return;
-    if(!user) return route.push('auth/login');
-    const collectionRef = collection(db, 'posts');
-    const q = query(collectionRef, where('user', '==', user.uid))
+    if (loading) return;
+    if (!user) return route.push("auth/login");
+    const collectionRef = collection(db, "posts");
+    const q = query(collectionRef, where("user", "==", user.uid));
     const unsubscribe = onSnapshot(q, (snapshot) => {
-      setPosts(snapshot.docs.map((doc) => ({...doc.data(), id:doc.id})))
-    })
+      const result = snapshot.docs.map((doc) => ({
+        ...doc.data(),
+        id: doc.id,
+      }));
+      setPosts(result);
+    });
     return unsubscribe;
-  }
+  };
 
   //Delete Post
-  const deletePost = async (id) => {
-    const docRef = doc(db, 'posts', id)
-    await deleteDoc(docRef)
-  }
+  const deletePost = async (id?: string) => {
+    if (id) {
+      const docRef = doc(db, "posts", id);
+      await deleteDoc(docRef);
+    }
+  };
 
   //Get users data
   useEffect(() => {
@@ -39,21 +59,34 @@ export default function Dashboard(){
     <div>
       <h1>Your posts</h1>
       <div>
-        {posts.map((post) => {
+        {posts.map((post: Post) => {
           return (
             <Message key={post.id} {...post}>
-              <div className='flex gap-4'>
-                <button onClick={(e) => deletePost(post.id)} className='text-pink-600 flex items-center gap-2 py-2 text-sm'><BsTrash2Fill className='text-2xl'/>Delete</button>
-                <Link href={{pathname: "/post", query: post}}>
-                  <button className='text-teal-600 flex items-center gap-2 py-2 text-sm'><AiFillEdit className='text-2xl'/>Edit</button>
+              <div className="flex gap-4">
+                <button
+                  onClick={(e) => deletePost(post.id)}
+                  className="text-pink-600 flex items-center gap-2 py-2 text-sm"
+                >
+                  <BsTrash2Fill className="text-2xl" />
+                  Delete
+                </button>
+                <Link href={{ pathname: "/post", query: JSON.stringify(post) }}>
+                  <button className="text-teal-600 flex items-center gap-2 py-2 text-sm">
+                    <AiFillEdit className="text-2xl" />
+                    Edit
+                  </button>
                 </Link>
               </div>
             </Message>
-          )
-        }
-        )}
+          );
+        })}
       </div>
-      <button className='font-semibold text-white bg-gray-800 py-2 px-4 my-6 rounded-lg' onClick={() => auth.signOut()}>Sign out</button>
+      <button
+        className="font-semibold text-white bg-gray-800 py-2 px-4 my-6 rounded-lg"
+        onClick={() => auth.signOut()}
+      >
+        Sign out
+      </button>
     </div>
-  )
+  );
 }
